Cookie bar shows when JavaScript turned off.
Since Django sets cookies server side, cookies are still set when
JavaScript is turned off.

So the cookie bar needs to be shown when JavaScript is turned off, but
this will cause jank is the cookie bar is shown and hidden with
JavaScript.

Solution uses the `.js-enabled` class to hide the bar by default, then
the bar is shown if JavaScript finds the `seen_cookie_bar` cookie.

* JS off - cookie bar is always shown.
* JS on - cookies are read; if 'seen_cookie_bar' cookie hasn't been set, then the bar is shown