Replace Django admin page with SSO integration
Note: this change will likely require rebuilding the DB because of a custom user model and the order that migrations are applied

When a user is authenticated via SSO their local django user is get-or-created. If a user is created, however, it will not have is_staff permissions and thus will not be able to access the admin section.  This needs to be manually granted via the admin.