Javascript compilation step for GOV.UK Frontend
In order to provide details component functionality for all
supported browsers, we need to include a polyfill from GOV.UK Frontend.

As we're not using a lot of the components, we want to selectively
import just the details and button javascript (for accessibility).

This PR adds browserify to import required javascript and a gulp task to
handle compilation.