Update measure type ids to be display for international duty group type
This is based on the "Measure types per measure type series" table at https://uktrade.github.io/tariff-data-manual/documentation/tariff-data-structures/measure-types.html#introduction-to-measure-types.

This is from the row "Series C - Applicable duty".

At time of this commit we are unsure whether it is required to update the rest of this list as this was to fix a specific problem where a tariff/charge wasn't displaying for a particular country on our end but was on the Trade Tariff side.