UKGT updated wording with links to new future tariff tool
Updates on UKGT descriptions to add more consistency, and also linking to the new tariff lookup tool where there are preferential agreements.

This change also adds a different template for Ecuador - ANDEAN-COUNTRIES_RATIFIED - which will need to be changed in the database. Once deployed a separate ticket can be created for this.

Technically ICELAND_NORWAY could be retired as it is now the same as EU_AGR_SIGNED_LINK. However this is probably not worth the effort given shelf-life over the overall feature set.