Mar2020 CNC updates - batch one
ONGOING-MFN-NOW template to apply to Algeria, Bosnia, Serbia, Montenegro (DZ, BA, RS, ME)

ONGOING-GSP-NOW template to apply to Ghana. Ghana has its own current template, but we will deprecate that for the more general case.

Amends to Canada
Amends to Japan

A separate Jira ticket will be made for the data amends.

There's going to be another PR for either a global change to the EU_AGR_SIGNED_LINK or a special case being created for ZA,KR,CH. It's unclear whether "priority updates" is about trying to save us work but accidentally creates it.